Support both Postgres and MySQL in the sqldb
* new query building methods in the sqldb
* rebind switches bindvars for flavor

[#119290575]

+++ db/sqldb/sqldb_suite_test.go
@@ -20,7 +20,7 @@ import (
 	"github.com/pivotal-golang/lager/lagertest"
 	"github.com/tedsuo/ifrit"
 
-	_ "github.com/go-sql-driver/mysql"
+	_ "github.com/lib/pq"
 
 	"testing"
 )
@@ -35,6 +35,7 @@ var (
 	serializer       format.Serializer
 	migrationProcess ifrit.Process
 	useSQL           bool
+	usePostgres      bool
 )
 
 func TestSql(t *testing.T) {
@@ -48,6 +49,7 @@ var _ = BeforeSuite(func() {
 	if !useSQL {
 		return
 	}
+	usePostgres = os.Getenv("USE_POSTGRES") == "true"
 
 	var err error
 	fakeClock = fakeclock.NewFakeClock(time.Now())
@@ -56,14 +58,22 @@ var _ = BeforeSuite(func() {
 
 	// mysql must be set up on localhost as described in the CONTRIBUTING.md doc
 	// in diego-release.
-	db, err = sql.Open("mysql", "diego:diego_password@/")
+	if usePostgres {
+		db, err = sql.Open("postgres", "postgres://diego:diego_pw@localhost")
+	} else {
+		db, err = sql.Open("mysql", "diego:diego_password@/")
+	}
 	Expect(err).NotTo(HaveOccurred())
 	Expect(db.Ping()).NotTo(HaveOccurred())
 
 	_, err = db.Exec(fmt.Sprintf("CREATE DATABASE diego_%d", GinkgoParallelNode()))
 	Expect(err).NotTo(HaveOccurred())
 
-	db, err = sql.Open("mysql", fmt.Sprintf("diego:diego_password@/diego_%d", GinkgoParallelNode()))
+	if usePostgres {
+		db, err = sql.Open("postgres", fmt.Sprintf("postgres://diego:diego_pw@localhost/diego_%d", GinkgoParallelNode()))
+	} else {
+		db, err = sql.Open("mysql", fmt.Sprintf("diego:diego_password@/diego_%d", GinkgoParallelNode()))
+	}
 	Expect(err).NotTo(HaveOccurred())
 	Expect(db.Ping()).NotTo(HaveOccurred())
 
@@ -74,7 +84,7 @@ var _ = BeforeSuite(func() {
 	cryptor = encryption.NewCryptor(keyManager, rand.Reader)
 	serializer = format.NewSerializer(cryptor)
 
-	internalSQLDB := sqldb.NewSQLDB(db, 5, 5, format.ENCRYPTED_PROTO, cryptor, fakeGUIDProvider, fakeClock)
+	internalSQLDB := sqldb.NewSQLDB(db, 5, 5, format.ENCRYPTED_PROTO, cryptor, fakeGUIDProvider, fakeClock, sqldb.Postgres)
 	err = internalSQLDB.CreateConfigurationsTable(logger)
 	if err != nil {
 		logger.Fatal("sql-failed-create-configurations-table", err)
@@ -116,10 +126,18 @@ var _ = AfterEach(func() {
 
 var _ = AfterSuite(func() {
 	if useSQL {
-		migrationProcess.Signal(os.Kill)
-		_, err := db.Exec(fmt.Sprintf("DROP DATABASE diego_%d", GinkgoParallelNode()))
+		if migrationProcess != nil {
+			migrationProcess.Signal(os.Kill)
+		}
+		var err error
+		if usePostgres {
+			Expect(db.Close()).NotTo(HaveOccurred())
+			db, err = sql.Open("postgres", "postgres://diego:diego_pw@localhost")
+			Expect(err).NotTo(HaveOccurred())
+			Expect(db.Ping()).NotTo(HaveOccurred())
+		}
+		_, err = db.Exec(fmt.Sprintf("DROP DATABASE diego_%d", GinkgoParallelNode()))
 		Expect(err).NotTo(HaveOccurred())
-
 		Expect(db.Close()).NotTo(HaveOccurred())
 	}
 })
